use boxing and unboxing of primitive JAVA values
git-svn-id: https://server.mycore.de/svn/mycore/trunk@25648 7089d063-bc14-496d-ac90-53de1872eb6c

+++ mycore-base/src/main/java/org/mycore/access/mcrimpl/MCRAccessControlSystem.java
@@ -333,13 +333,13 @@ public class MCRAccessControlSystem extends MCRAccessBaseImpl {
         int nextFreeID;
         String sNextFreeID;
         if (nextFreeRuleID.containsKey(prefix)) {
-            nextFreeID = nextFreeRuleID.get(prefix).intValue();
+            nextFreeID = nextFreeRuleID.get(prefix);
         } else {
             nextFreeID = ruleStore.getNextFreeRuleID(prefix);
         }
         sNextFreeID = lexicographicalPattern + String.valueOf(nextFreeID);
         sNextFreeID = sNextFreeID.substring(sNextFreeID.length() - lexicographicalPattern.length());
-        nextFreeRuleID.put(prefix, Integer.valueOf(nextFreeID + 1));
+        nextFreeRuleID.put(prefix, nextFreeID + 1);
         return prefix + sNextFreeID;
     }
